reverting to the original version of the codebase
although i did want to rewrite the app during my gap year, i eventually realised that keeping it the way it was when i wrote it in 11th is far more fitting than anything i could do with it now.

quite frankly, the app serves to practical purpose, and rewriting it from the ground up now is both a waste of my time as well as a loss of a historical record in the way my programming has changed.

thus, i'm going to revert it to an untouched state from when i last left off on it a few years ago. i might however come back and touch up a lot of the code to make sure it actually compiles properly on people's machines, since i do think it has a nasty history of breaking all the time.